Updated Cart Features in Cart Page also

+++ lib/Screens/home/food_page_body.dart
@@ -60,7 +60,7 @@ class _FoodPageBodyState extends State<FoodPageBody> {
             dotsCount: popularProduct.popularProductList.isEmpty
                 ? 1
                 : popularProduct.popularProductList.length,
-            position: currPageValue.toDouble(),
+            position: currPageValue,
             decorator: DotsDecorator(
               activeColor: AppColors.mainColor,
               size: const Size.square(9.0),
